requirements.txt
Django>=4.0.1, <4.1
djangorestframework>=3.13.1,<3.14
psycopg2>2.9.3,<2.10
django-phonenumber-field
phonenumbers
drf-spectacular>=0.22.1,<0.23
django-cors-headers
Pillow>=9.1.0, <9.2
uwsgi>=2.0.20, <2.1
django-filter
ip2geotools>=0.1.5,<0.1.6
firebase-admin
django-imagekit
channels
daphne

